LastAction wrapper corrected

+++ tests/test_models.py
@@ -5,17 +5,17 @@ import numpy as np
 
 def test_obs_eq():
     obs1 = Observation(
-        data=np.arange(20),
+        data=np.arange(20, dtype=np.float32),
         available_actions=np.ones(5, dtype=np.float32),
         state=np.ones(10, dtype=np.float32),
-        extras=np.arange(5),
+        extras=np.arange(5, dtype=np.float32),
     )
 
     obs2 = Observation(
-        data=np.arange(20),
+        data=np.arange(20, dtype=np.float32),
         available_actions=np.ones(5, dtype=np.float32),
         state=np.ones(10, dtype=np.float32),
-        extras=np.arange(5),
+        extras=np.arange(5, dtype=np.float32),
     )
 
     assert obs1 == obs2
@@ -23,14 +23,14 @@ def test_obs_eq():
 
 def test_obs_eq_extras_none():
     obs1 = Observation(
-        data=np.arange(20),
+        data=np.arange(20, dtype=np.float32),
         available_actions=np.ones(5, dtype=np.float32),
         state=np.ones(10, dtype=np.float32),
         extras=None,
     )
 
     obs2 = Observation(
-        data=np.arange(20),
+        data=np.arange(20, dtype=np.float32),
         available_actions=np.ones(5, dtype=np.float32),
         state=np.ones(10, dtype=np.float32),
         extras=None,
@@ -41,17 +41,17 @@ def test_obs_eq_extras_none():
 
 def test_obs_hash():
     obs1 = Observation(
-        data=np.arange(20),
+        data=np.arange(20, dtype=np.float32),
         available_actions=np.ones(5, dtype=np.float32),
         state=np.ones(10, dtype=np.float32),
-        extras=np.arange(5),
+        extras=np.arange(5, dtype=np.float32),
     )
 
     obs2 = Observation(
-        data=np.arange(20),
+        data=np.arange(20, dtype=np.float32),
         available_actions=np.ones(5, dtype=np.float32),
         state=np.ones(10, dtype=np.float32),
-        extras=np.arange(5),
+        extras=np.arange(5, dtype=np.float32),
     )
 
     assert hash(obs1) == hash(obs2)
@@ -60,40 +60,40 @@ def test_obs_hash():
 def test_transition_eq():
     t1 = Transition(
         obs=Observation(
-            data=np.arange(20),
+            data=np.arange(20, dtype=np.float32),
             available_actions=np.ones(5, dtype=np.float32),
             state=np.ones(10, dtype=np.float32),
-            extras=np.arange(5),
+            extras=np.arange(5, dtype=np.float32),
         ),
         action=np.ones(5, dtype=np.float32),
         reward=1.0,
         done=False,
         info={},
         obs_=Observation(
-            data=np.arange(20),
+            data=np.arange(20, dtype=np.float32),
             available_actions=np.ones(5, dtype=np.float32),
             state=np.ones(10, dtype=np.float32),
-            extras=np.arange(5),
+            extras=np.arange(5, dtype=np.float32),
         ),
         truncated=False,
     )
 
     t2 = Transition(
         obs=Observation(
-            data=np.arange(20),
+            data=np.arange(20, dtype=np.float32),
             available_actions=np.ones(5, dtype=np.float32),
             state=np.ones(10, dtype=np.float32),
-            extras=np.arange(5),
+            extras=np.arange(5, dtype=np.float32),
         ),
         action=np.ones(5, dtype=np.float32),
         reward=1.0,
         done=False,
         info={},
         obs_=Observation(
-            data=np.arange(20),
+            data=np.arange(20, dtype=np.float32),
             available_actions=np.ones(5, dtype=np.float32),
             state=np.ones(10, dtype=np.float32),
-            extras=np.arange(5),
+            extras=np.arange(5, dtype=np.float32),
         ),
         truncated=False,
     )
@@ -104,40 +104,40 @@ def test_transition_eq():
 def test_transition_hash():
     t1 = Transition(
         obs=Observation(
-            data=np.arange(20),
+            data=np.arange(20, dtype=np.float32),
             available_actions=np.ones(5, dtype=np.float32),
             state=np.ones(10, dtype=np.float32),
-            extras=np.arange(5),
+            extras=np.arange(5, dtype=np.float32),
         ),
         action=np.ones(5, dtype=np.float32),
         reward=1.0,
         done=False,
         info={},
         obs_=Observation(
-            data=np.arange(20),
+            data=np.arange(20, dtype=np.float32),
             available_actions=np.ones(5, dtype=np.float32),
             state=np.ones(10, dtype=np.float32),
-            extras=np.arange(5),
+            extras=np.arange(5, dtype=np.float32),
         ),
         truncated=False,
     )
 
     t2 = Transition(
         obs=Observation(
-            data=np.arange(20),
+            data=np.arange(20, dtype=np.float32),
             available_actions=np.ones(5, dtype=np.float32),
             state=np.ones(10, dtype=np.float32),
-            extras=np.arange(5),
+            extras=np.arange(5, dtype=np.float32),
         ),
         action=np.ones(5, dtype=np.float32),
         reward=1.0,
         done=False,
         info={},
         obs_=Observation(
-            data=np.arange(20),
+            data=np.arange(20, dtype=np.float32),
             available_actions=np.ones(5, dtype=np.float32),
             state=np.ones(10, dtype=np.float32),
-            extras=np.arange(5),
+            extras=np.arange(5, dtype=np.float32),
         ),
         truncated=False,
     )
